♻️ refactor: makeParams file ts 변환 #251
